requirements.txt
djangorestframework-simplejwt
Django~=5.2.6
pip~=25.2
mysqlclient
djangorestframework
django-cors-headers

